Fix clientIndex range check for game.maxClients
Updated the conditional check for clientIndex to cast
game.maxClients to an int using static_cast<int>. This
ensures proper comparison between signed integers,
preventing potential issues with negative values and
ensuring correct behavior with large client counts.

+++ src/command_system.cpp
@@ -106,7 +106,7 @@ namespace Commands {
 
                 for (auto p_ent : active_players()) {
                         const int clientIndex = p_ent->s.number - 1;
-                        if (clientIndex < 0 || clientIndex >= game.maxClients) {
+                        if (clientIndex < 0 || clientIndex >= static_cast<int>(game.maxClients)) {
                                 continue;
                         }
